Move functions into layers.ts and data.ts
