+++ web/src/data.ts
@@ -1,4 +1,15 @@
-import { Dataset, DerivedRegion } from "./types";
+import { Dataset, DerivedRegion } from "@/types";
+import {
+  currentMapDataSource,
+  activeMapLayerIds,
+  selectedDataSourceIds,
+} from "@/store";
+import { getUid } from "ol/util";
+import {
+  addDataSourceLayerToMap,
+  getMapLayerFromDataSource,
+  updateVisibleLayers,
+} from "@/layers";
 
 interface MapDataSourceArgs {
   dataset?: Dataset;
@@ -37,5 +48,48 @@ export class MapDataSource {
     }
 
     return name;
+  }
+}
+
+export function addDataSourceToMap(dataSource: MapDataSource) {
+  // Add dataset id to selected datasets
+  selectedDataSourceIds.add(dataSource.getUid());
+
+  // Check if layer with this dataset already exists
+  const existingLayer = getMapLayerFromDataSource(dataSource);
+
+  // Get either existing layer or create a new one
+  const layer = existingLayer || addDataSourceLayerToMap(dataSource);
+  if (layer === undefined) {
+    throw new Error("No layer returned when adding data source to map");
+  }
+
+  // Put new dataset at front of list, so it shows up above any existing layers
+  activeMapLayerIds.value = [getUid(layer), ...activeMapLayerIds.value];
+
+  // Re-order layers
+  updateVisibleLayers();
+}
+
+export function hideDataSourceFromMap(dataSource: MapDataSource) {
+  const dataSourceId = dataSource.getUid();
+  // Remove dataset id from selected datasets
+  selectedDataSourceIds.delete(dataSourceId);
+
+  // Filter out dataset layer from active map layers
+  const layer = getMapLayerFromDataSource(dataSource);
+  if (layer === undefined) {
+    throw new Error(`Couldn't find layer for data source ${dataSourceId}`);
+  }
+  activeMapLayerIds.value = activeMapLayerIds.value.filter(
+    (layerId) => layerId !== getUid(layer)
+  );
+
+  // Re-order layers
+  updateVisibleLayers();
+
+  // If current data source was the de-selected dataset, un-set it
+  if (currentMapDataSource.value?.getUid() === dataSourceId) {
+    currentMapDataSource.value = undefined;
   }
 }
