+++ web/src/utils.js
@@ -1,21 +1,8 @@
-import TileLayer from "ol/layer/Tile.js";
-import VectorLayer from "ol/layer/Vector";
-import VectorTileLayer from "ol/layer/VectorTile.js";
-import XYZSource from "ol/source/XYZ.js";
-import VectorSource from "ol/source/Vector";
-import VectorTileSource from "ol/source/VectorTile.js";
-import GeoJSON from "ol/format/GeoJSON.js";
 import { Fill, Stroke, Circle, Style } from "ol/style.js";
-import { Feature } from "ol";
-import { getUid } from "ol/util";
-import { LineString, Point } from "ol/geom";
-import { fromLonLat } from "ol/proj";
 
-import { baseURL } from "@/api/auth";
 import { getNetworkGCC, getCityCharts, getRasterData } from "@/api/rest";
 import {
   map,
-  showMapBaseLayer,
   currentCity,
   rasterTooltip,
   networkVis,
@@ -23,8 +10,6 @@ import {
   currentNetworkGCC,
   availableCharts,
   activeChart,
-  activeMapLayerIds,
-  selectedDataSourceIds,
 } from "@/store";
 
 export const rasterColormaps = [
@@ -51,65 +36,6 @@ export const rasterColormaps = [
 
 var rasterTooltipDataCache = {};
 
-/** Returns if a layer should be enabled based on showMapBaseLayer, activeMapLayerIds, and networkVis */
-function getLayerEnabled(layer) {
-  // Check if layer is map base layer
-  if (layer.getProperties().baseLayer) {
-    return showMapBaseLayer.value;
-  }
-
-  // Check if layer is enabled
-  const layerId = getUid(layer);
-  let layerEnabled = activeMapLayerIds.value.includes(layerId);
-  // TODO: Remove datasetId
-  if (networkVis.value) {
-    const layerDatasetId = layer.getProperties().datasetId;
-    if (layerDatasetId === networkVis.value.id) {
-      layerEnabled = layerEnabled && layer.getProperties().network;
-    }
-  }
-
-  return layerEnabled;
-}
-
-/**
- * Shows/hides layers based on the getLayerEnabled function.
- *
- * Note: This only modifies layer visibility. It does not actually enable or disable any map layers directly.
- * */
-export function updateVisibleLayers() {
-  const layerState = {
-    shown: [],
-    hidden: [],
-  };
-  const allLayers = map.value?.getLayers()?.getArray();
-  if (!allLayers) {
-    return layerState;
-  }
-
-  allLayers.forEach((layer) => {
-    const layerEnabled = getLayerEnabled(layer);
-    if (!layerEnabled) {
-      layer.setVisible(false);
-      layerState.hidden.push(layer);
-      return;
-    }
-
-    // Set layer visible and z-index
-    layer.setVisible(true);
-    layerState.shown.push(layer);
-    const layerId = getUid(layer);
-    const layerIndex = activeMapLayerIds.value.findIndex(
-      (id) => id === layerId
-    );
-    layer.setZIndex(
-      layerIndex > -1 ? activeMapLayerIds.value.length - layerIndex : 0
-    );
-  });
-
-  return layerState;
-}
-
 export function cacheRasterData(datasetId) {
   if (!rasterTooltipDataCache[datasetId]) {
     rasterTooltipDataCache[datasetId] = {};
@@ -119,7 +45,7 @@ export function cacheRasterData(datasetId) {
   }
 }
 
-function createStyle(args) {
+export function createStyle(args) {
   let colors = ["#00000022"];
   if (args.colors) {
     colors = args.colors.split(",");
@@ -162,112 +88,6 @@ function createStyle(args) {
       return new Style(styleSpec);
     });
   }
-}
-
-export function addDerivedRegionLayerToMap(region) {
-  const layer = new VectorLayer({
-    style: (feature) =>
-      createStyle({
-        type: feature.getGeometry().getType(),
-      }),
-    source: new VectorSource({
-      format: new GeoJSON(),
-      url: `${baseURL}derived_regions/${region.id}/boundary/`,
-    }),
-  });
-
-  // Add to map
-  map.value.addLayer(layer);
-  return layer;
-}
-
-export function addDatasetLayerToMap(dataset, zIndex) {
-  if (dataset.processing) {
-    return;
-  }
-  let layer;
-
-  // Add raster data
-  if (dataset.raster_file) {
-    const tileParams = {
-      projection: "EPSG:3857",
-      band: 1,
-      palette: dataset.style?.colormap || "terrain",
-    };
-    if (
-      dataset.style?.colormap_range !== undefined &&
-      dataset.style?.colormap_range.length === 2
-    ) {
-      tileParams.min = dataset.style.colormap_range[0];
-      tileParams.max = dataset.style.colormap_range[1];
-    }
-    if (dataset.style?.options?.transparency_threshold !== undefined) {
-      tileParams.nodata = dataset.style.options.transparency_threshold;
-    }
-    const tileParamString = Object.keys(tileParams)
-      .map((key) => key + "=" + tileParams[key])
-      .join("&");
-    layer = new TileLayer({
-      source: new XYZSource({
-        url: `${baseURL}datasets/${dataset.id}/tiles/{z}/{x}/{y}.png/?${tileParamString}`,
-      }),
-      opacity: dataset.style?.opacity || 1,
-      zIndex,
-    });
-
-    cacheRasterData(dataset.id);
-  } else if (dataset.geodata_file) {
-    // Use VectorLayer if dataset category is "region"
-    if (dataset.category === "region") {
-      layer = new VectorLayer({
-        zIndex,
-        style: (feature) =>
-          createStyle({
-            type: feature.getGeometry().getType(),
-            colors: feature.get("properties").colors,
-          }),
-        source: new VectorSource({
-          format: new GeoJSON(),
-          url: `${baseURL}datasets/${dataset.id}/regions`,
-        }),
-      });
-    } else {
-      // Use vector tiles
-      layer = new VectorTileLayer({
-        source: new VectorTileSource({
-          format: new GeoJSON(),
-          url: `${baseURL}datasets/${dataset.id}/vector-tiles/{z}/{x}/{y}/`,
-        }),
-        style: (feature) =>
-          createStyle({
-            type: feature.getGeometry().getType(),
-            colors: feature.get("colors"),
-          }),
-        opacity: dataset.style?.opacity || 1,
-        zIndex,
-      });
-    }
-  }
-
-  // Add to map
-  map.value.addLayer(layer);
-  return layer;
-}
-
-export function addDataSourceLayerToMap(dataSource) {
-  let layer;
-  if (dataSource.dataset) {
-    layer = addDatasetLayerToMap(
-      dataSource.dataset,
-      selectedDataSourceIds.size - 1
-    );
-  } else if (dataSource.derivedRegion) {
-    layer = addDerivedRegionLayerToMap(dataSource.derivedRegion);
-  }
-
-  // Add this to link layers to data sources
-  layer.setProperties({ dataSourceId: dataSource.getUid() });
-  return layer;
 }
 
 function getNetworkFeatureStyle(alpha = "ff", highlight = false) {
@@ -363,53 +183,6 @@ export function updateNetworkStyle() {
         });
       }
     });
-}
-
-// TODO: Roll into addDatasetLayerToMap
-export function addNetworkLayerToMap(dataset, nodes) {
-  const source = new VectorSource();
-  const features = [];
-  const visitedNodes = [];
-  nodes.forEach((node) => {
-    features.push(
-      new Feature(
-        Object.assign(node.properties, {
-          name: node.name,
-          id: node.id,
-          node: true,
-          geometry: new Point(fromLonLat(node.location.toReversed())),
-        })
-      )
-    );
-    node.adjacent_nodes.forEach((adjId) => {
-      if (!visitedNodes.includes(adjId)) {
-        const adjNode = nodes.find((n) => n.id === adjId);
-        features.push(
-          new Feature({
-            connects: [node.id, adjId],
-            edge: true,
-            geometry: new LineString([
-              fromLonLat(node.location.toReversed()),
-              fromLonLat(adjNode.location.toReversed()),
-            ]),
-          })
-        );
-      }
-    });
-    visitedNodes.push(node.id);
-  });
-  source.addFeatures(features);
-
-  const layer = new VectorLayer({
-    properties: {
-      network: true,
-    },
-    zIndex: 99,
-    style: getNetworkFeatureStyle(),
-    source,
-  });
-  map.value.addLayer(layer);
-  return layer;
 }
 
 export function displayRasterTooltip(evt, tooltip, overlay) {
